Changed quota from int to long to accomodate more reasonable quota sizes
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/WDK/trunk@21820 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ Model/src/main/java/org/gusdb/wdk/model/user/dataset/UserDatasetInfo.java
@@ -26,7 +26,7 @@ public class UserDatasetInfo {
   private final UserDataset _userDataset;
   private final boolean _isInstalled;
   private final User _owner;
-  private final int _ownerQuota;
+  private final long _ownerQuota;
   private final List<String> _relevantQuestionNames;
   private final List<UserDatasetShareUser> _shares;
   private final UserDatasetTypeHandler _handler;
@@ -89,7 +89,7 @@ public class UserDatasetInfo {
     return _owner;
   }
 
-  public int getOwnerQuota() {
+  public long getOwnerQuota() {
     return _ownerQuota;
   }
